requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
beautifulsoup4==4.11.1
fastapi==0.86.0
fuzzywuzzy==0.18.0
matplotlib==3.4.3
python_dateutil==2.8.2
requests==2.24.0
